requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aerich==0.7.2
    # via core-sim-3 (pyproject.toml)
aiosqlite==0.17.0
    # via tortoise-orm
annotated-types==0.7.0
    # via
    #   core-sim-3 (pyproject.toml)
    #   pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.29.0
    # via tortoise-orm
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   aerich
    #   typer
    #   uvicorn
dictdiffer==0.9.0
    # via aerich
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.112.2
    # via core-sim-3 (pyproject.toml)
fastapi-cli==0.0.5
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.8
    # via
    #   anyio
    #   email-validator
    #   httpx
iso8601==1.1.0
    # via tortoise-orm
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
newrelic==9.13.0
    # via core-sim-3 (pyproject.toml)
pydantic==2.8.2
    # via
    #   aerich
    #   fastapi
    #   tortoise-orm
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
pypika-tortoise==0.1.6
    # via tortoise-orm
python-dotenv==1.0.1
    # via
    #   core-sim-3 (pyproject.toml)
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via tortoise-orm
pyyaml==6.0.2
    # via uvicorn
rich==13.8.0
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.4
    # via fastapi
tomlkit==0.13.2
    # via aerich
tortoise-orm==0.21.6
    # via
    #   core-sim-3 (pyproject.toml)
    #   aerich
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
ulid-py==1.1.0
    # via core-sim-3 (pyproject.toml)
uvicorn==0.30.6
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
